refactor: [#168] integrate progress module into views layer
Move src/presentation/progress.rs to src/presentation/views/progress/mod.rs
to consolidate all user-facing output components under the views layer.

Updated import paths across 7 files:
- 6 controller files (create, destroy subcommands)
- 1 test file (reentrancy test)

All documentation examples updated to reflect new module structure.
Tests and linting pass. No breaking changes to public API.
